+++ app/src/main/java/com/shopapp/lamaria/models/Product.java
@@ -6,7 +6,7 @@ public class Product {
     private String product_name;
     private String category_id;
     private String category_name;
-    private double product_price;
+    private int product_price;
     private String product_status;
     private String product_image;
     private String product_description;
@@ -30,7 +30,7 @@ public class Product {
         return category_name;
     }
 
-    public double getProduct_price() {
+    public int getProduct_price() {
         return product_price;
     }
